requirements.txt
pyspark==3.4.0
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.1
seaborn==0.12.2
jupyter==1.0.0
PyYAML==6.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
pytest==7.4.0
black==23.7.0
flake8==6.0.0
